% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{Get Default Attributes}
\usage{
get_attributes()
}
\value{
list of default general and variable attribute data frames. The data
frames include the attribute names and a description of the sort of
information that should be included in the metadata.
}
\description{
Exposes the default general and variable attributes for writing metadata
template files.
}
\examples{
a <- get_attributes()
a$gen_attributes
a$var_attributes
}
