% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_meta_template.R
\name{write_meta_template}
\alias{write_meta_template}
\title{Write csv file with metadata template}
\usage{
write_meta_template(
  dData,
  file,
  embedded = TRUE,
  gen_attributes = gen_attrib$attribute_name,
  var_attributes = var_attrib$attribute_name
)
}
\arguments{
\item{dData}{Data frame to be written as csv file with metadata template}

\item{file}{Base name of output csv file}

\item{embedded}{Are the metada added to the csv file above the data or is the
metadata written into a seperate file. If the data are embedded, the output
file name is "file_withMeta.csv". If the data are not embedded, the data
file is named "file.csv" and the metadate file is named "file_meta.csv".
Default embedded = TRUE}

\item{gen_attributes}{Vector of general attributes for metadata template.
Default is gen_format_2 from the internal list of format options. Can be
any user provided character vector that does not contain any commas.}

\item{var_attributes}{Vector of variable attributes for metadata template.
Default is gen_format_2 from the internal list of format options. Can be
any user provided character vector that does not contain any commas.}
}
\value{
Writes csv file(s).
}
\description{
Write csv file with metadata template
}
\examples{
write_meta_template(iris, "~/iris")
write_meta_template(iris, "~/iris", embedded = FALSE)
}
