% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta.R
\name{read_meta}
\alias{read_meta}
\title{Read csv file with metadata}
\usage{
read_meta(file)
}
\arguments{
\item{file}{path and name of the .csv data file to read}
}
\value{
List with three data frames
\describe{
\item{dGeneralMeta}{General metadata with attribute and value}
\item{dVariableMeta}{Table of metadata for each dataset variable}
\item{dData}{The dataset}
}
}
\description{
Read csv file with metadata
}
\examples{
iris_withMeta_raw <- read.csv(system.file("iris_withMeta.csv", package = "metacsv" ))
iris_withMeta_parsed <- read_meta(system.file("iris_withMeta.csv", package = "metacsv" ))
}
